requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
anyio==4.12.1
apache-airflow==2.5.1
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==3.3.2
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==3.3.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.4
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.8.0
graphviz==0.21
greenlet==3.3.1
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgspec==0.20.0
multidict==6.7.1
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pathspec==0.9.0
pendulum==3.2.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
setproctitle==1.3.7
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.52.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
WTForms==3.2.1
yarl==1.22.0
